Fix a bug with not extending redis session cookies
There was a bug with the way oz was dealing with redis session. Oz would
extend the session key in redis, but didn't extend the expiration of the
`session_id`. This would cause all users to be logged out after the
cookie expiration time expired (which would be 30 days since the value
wasn't being set by oz). This update ensure that cookie is extended and
reset everytime the session key is accessed.